+++ Question/package.bluej
@@ -12,199 +12,212 @@ objectbench.height=100
 objectbench.width=849
 package.editor.height=605
 package.editor.width=751
-package.editor.x=-60
-package.editor.y=38
+package.editor.x=586
+package.editor.y=220
 package.numDependencies=3
-package.numTargets=15
+package.numTargets=16
 package.showExtends=true
 package.showUses=true
 project.charset=US-ASCII
-target1.editor.height=632
+target1.editor.height=700
 target1.editor.width=900
-target1.editor.x=576
-target1.editor.y=181
+target1.editor.x=452
+target1.editor.y=285
 target1.height=50
-target1.name=Billfold
+target1.name=TriangleFrame
 target1.naviview.expanded=true
 target1.showInterface=false
 target1.type=ClassTarget
 target1.typeParameters=
-target1.width=80
-target1.x=410
-target1.y=180
+target1.width=110
+target1.x=20
+target1.y=380
 target10.editor.height=700
 target10.editor.width=900
-target10.editor.x=414
-target10.editor.y=463
+target10.editor.x=318
+target10.editor.y=317
 target10.height=50
-target10.name=NumericQuestion
+target10.name=Coin
 target10.naviview.expanded=true
 target10.showInterface=false
 target10.type=ClassTarget
 target10.typeParameters=
-target10.width=120
-target10.x=200
-target10.y=0
+target10.width=80
+target10.x=320
+target10.y=120
 target11.editor.height=700
 target11.editor.width=900
-target11.editor.x=391
-target11.editor.y=298
+target11.editor.x=414
+target11.editor.y=463
 target11.height=50
-target11.name=ChoiceQuestion
+target11.name=NumericQuestion
 target11.naviview.expanded=true
 target11.showInterface=false
 target11.type=ClassTarget
 target11.typeParameters=
-target11.width=110
-target11.x=180
-target11.y=190
+target11.width=120
+target11.x=200
+target11.y=0
 target12.editor.height=700
 target12.editor.width=900
-target12.editor.x=690
-target12.editor.y=190
+target12.editor.x=391
+target12.editor.y=298
 target12.height=50
-target12.name=Question
+target12.name=ChoiceQuestion
 target12.naviview.expanded=true
 target12.showInterface=false
 target12.type=ClassTarget
 target12.typeParameters=
-target12.width=80
-target12.x=160
-target12.y=100
+target12.width=110
+target12.x=180
+target12.y=190
+target13.editor.height=700
+target13.editor.width=900
+target13.editor.x=690
+target13.editor.y=190
 target13.height=50
-target13.name=QuestionDemo1
+target13.name=Question
+target13.naviview.expanded=true
 target13.showInterface=false
 target13.type=ClassTarget
 target13.typeParameters=
-target13.width=130
-target13.x=30
-target13.y=190
-target14.editor.height=786
-target14.editor.width=900
-target14.editor.x=210
-target14.editor.y=372
+target13.width=80
+target13.x=160
+target13.y=100
 target14.height=50
-target14.name=Test_person
-target14.naviview.expanded=true
+target14.name=QuestionDemo1
 target14.showInterface=false
 target14.type=ClassTarget
 target14.typeParameters=
-target14.width=100
-target14.x=0
-target14.y=520
-target15.editor.height=700
+target14.width=130
+target14.x=30
+target14.y=190
+target15.editor.height=786
 target15.editor.width=900
-target15.editor.x=172
-target15.editor.y=321
+target15.editor.x=210
+target15.editor.y=372
 target15.height=50
-target15.name=Person
+target15.name=Test_person
 target15.naviview.expanded=true
 target15.showInterface=false
 target15.type=ClassTarget
 target15.typeParameters=
-target15.width=80
-target15.x=100
-target15.y=530
-target2.editor.height=700
+target15.width=100
+target15.x=0
+target15.y=520
+target16.editor.height=700
+target16.editor.width=900
+target16.editor.x=172
+target16.editor.y=321
+target16.height=50
+target16.name=Person
+target16.naviview.expanded=true
+target16.showInterface=false
+target16.type=ClassTarget
+target16.typeParameters=
+target16.width=80
+target16.x=100
+target16.y=530
+target2.editor.height=632
 target2.editor.width=900
-target2.editor.x=179
-target2.editor.y=59
+target2.editor.x=576
+target2.editor.y=181
 target2.height=50
-target2.name=IDCard
+target2.name=Billfold
 target2.naviview.expanded=true
 target2.showInterface=false
 target2.type=ClassTarget
 target2.typeParameters=
 target2.width=80
-target2.x=550
-target2.y=310
+target2.x=410
+target2.y=180
 target3.editor.height=700
 target3.editor.width=900
-target3.editor.x=297
-target3.editor.y=306
+target3.editor.x=179
+target3.editor.y=59
 target3.height=50
-target3.name=DriverLicense
+target3.name=IDCard
 target3.naviview.expanded=true
 target3.showInterface=false
 target3.type=ClassTarget
 target3.typeParameters=
-target3.width=100
-target3.x=220
+target3.width=80
+target3.x=550
 target3.y=310
 target4.editor.height=700
 target4.editor.width=900
-target4.editor.x=0
-target4.editor.y=161
+target4.editor.x=297
+target4.editor.y=306
 target4.height=50
-target4.name=ButtonViewer
+target4.name=DriverLicense
 target4.naviview.expanded=true
 target4.showInterface=false
 target4.type=ClassTarget
 target4.typeParameters=
 target4.width=100
-target4.x=340
-target4.y=40
-target5.editor.height=809
+target4.x=220
+target4.y=310
+target5.editor.height=700
 target5.editor.width=900
-target5.editor.x=167
-target5.editor.y=647
+target5.editor.x=0
+target5.editor.y=161
 target5.height=50
-target5.name=Triangle
+target5.name=ButtonViewer
 target5.naviview.expanded=true
 target5.showInterface=false
 target5.type=ClassTarget
 target5.typeParameters=
-target5.width=80
-target5.x=60
-target5.y=290
-target6.editor.height=700
+target5.width=100
+target5.x=340
+target5.y=40
+target6.editor.height=809
 target6.editor.width=900
-target6.editor.x=441
-target6.editor.y=46
+target6.editor.x=153
+target6.editor.y=157
 target6.height=50
-target6.name=FillInQuestion
+target6.name=Triangle
+target6.naviview.expanded=true
 target6.showInterface=false
 target6.type=ClassTarget
 target6.typeParameters=
-target6.width=100
-target6.x=80
-target6.y=0
+target6.width=80
+target6.x=60
+target6.y=290
 target7.editor.height=700
 target7.editor.width=900
-target7.editor.x=727
-target7.editor.y=353
+target7.editor.x=441
+target7.editor.y=46
 target7.height=50
-target7.name=CallingCard
-target7.naviview.expanded=true
+target7.name=FillInQuestion
 target7.showInterface=false
 target7.type=ClassTarget
 target7.typeParameters=
-target7.width=90
-target7.x=400
-target7.y=430
+target7.width=100
+target7.x=80
+target7.y=0
 target8.editor.height=700
 target8.editor.width=900
-target8.editor.x=21
-target8.editor.y=61
+target8.editor.x=727
+target8.editor.y=353
 target8.height=50
-target8.name=Card
+target8.name=CallingCard
 target8.naviview.expanded=true
 target8.showInterface=false
 target8.type=ClassTarget
 target8.typeParameters=
-target8.width=80
+target8.width=90
 target8.x=400
-target8.y=310
+target8.y=430
 target9.editor.height=700
 target9.editor.width=900
-target9.editor.x=318
-target9.editor.y=317
+target9.editor.x=21
+target9.editor.y=61
 target9.height=50
-target9.name=Coin
+target9.name=Card
 target9.naviview.expanded=true
 target9.showInterface=false
 target9.type=ClassTarget
 target9.typeParameters=
 target9.width=80
-target9.x=320
-target9.y=120
+target9.x=400
+target9.y=310
